# public/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Reading Advantage LLM Benchmark API",
    "description": "Public read-only API for accessing benchmark data including leaderboard, runs, and tasks.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://benchmark.readingadvantage.com",
      "description": "Production server"
    }
  ],
  "paths": {
    "/api/v1/leaderboard": {
      "get": {
        "summary": "Get leaderboard",
        "description": "Returns ranked models with aggregated scores from all runs.",
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "description": "Filter by provider (e.g., openai, anthropic)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "leaderboard": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "model": { "type": "string" },
                          "runs": { "type": "integer" },
                          "latestRunDate": { "type": "string" },
                          "meanScore": { "type": "number" },
                          "medianScore": { "type": "number" },
                          "p95Score": { "type": "number" }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid or missing API key" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
    "/api/v1/runs": {
      "get": {
        "summary": "List runs",
        "description": "Returns paginated list of benchmark runs with optional filtering.",
        "parameters": [
          {
            "name": "model",
            "in": "query",
            "description": "Filter by model name",
            "schema": { "type": "string" }
          },
          {
            "name": "task",
            "in": "query",
            "description": "Filter by task ID",
            "schema": { "type": "string" }
          },
          {
            "name": "minScore",
            "in": "query",
            "description": "Minimum total score (0-1 scale)",
            "schema": { "type": "number" }
          },
          {
            "name": "maxScore",
            "in": "query",
            "description": "Maximum total score (0-1 scale)",
            "schema": { "type": "number" }
          },
          {
            "name": "startDate",
            "in": "query",
            "description": "Filter runs after this date (ISO format)",
            "schema": { "type": "string" }
          },
          {
            "name": "endDate",
            "in": "query",
            "description": "Filter runs before this date (ISO format)",
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of results to return",
            "schema": { "type": "integer", "default": 100 }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "Runs list response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "total": { "type": "integer" },
                    "limit": { "type": "integer" },
                    "offset": { "type": "integer" },
                    "runs": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Run" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runs/{id}": {
      "get": {
        "summary": "Get run by ID",
        "description": "Returns full details of a specific benchmark run.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Run ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Run details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Run" }
              }
            }
          },
          "404": { "description": "Run not found" }
        }
      }
    },
    "/api/v1/tasks": {
      "get": {
        "summary": "List tasks",
        "description": "Returns catalog of benchmark tasks with optional filtering.",
        "parameters": [
          {
            "name": "difficulty",
            "in": "query",
            "description": "Filter by difficulty level",
            "schema": { "type": "string", "enum": ["easy", "medium", "hard"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Tasks list response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "tasks": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Task" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{id}": {
      "get": {
        "summary": "Get task by ID",
        "description": "Returns full details of a specific benchmark task.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task ID",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Task details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Task" }
              }
            }
          },
          "404": { "description": "Task not found" }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Run": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "model": { "type": "string" },
          "provider": { "type": "string" },
          "harness": { "type": "string" },
          "harness_version": { "type": "string" },
          "benchmark_version": { "type": "string" },
          "run_date": { "type": "string" },
          "wall_time_seconds": { "type": "number" },
          "total_score": { "type": "number" },
          "scores": {
            "type": "object",
            "properties": {
              "functional_correctness": { "type": "number" },
              "integration_quality": { "type": "number" },
              "regression_safety": { "type": "number" },
              "minimality": { "type": "number" },
              "process_quality": { "type": "number" }
            }
          },
          "test_results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "suite": { "type": "string" },
                "name": { "type": "string" },
                "status": { "type": "string", "enum": ["pass", "fail", "skip"] },
                "duration_ms": { "type": "number" }
              }
            }
          },
          "artifacts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": { "type": "string" },
                "type": { "type": "string", "enum": ["log", "screenshot", "report", "other"] },
                "url": { "type": "string" }
              }
            }
          }
        }
      },
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "title": { "type": "string" },
          "difficulty": { "type": "string", "enum": ["easy", "medium", "hard"] },
          "domain": { "type": "string" },
          "description": { "type": "string" },
          "version": { "type": "string" },
          "status": { "type": "string", "enum": ["draft", "review", "published"] }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication"
      }
    }
  },
  "security": [{ "ApiKeyAuth": [] }]
}